nexed/verzamelen/02-level/Van-Archief-naar-Database-5f111952835d-eae111c5d246/requirements.txt
mysql-connector-python==9.3.0
pandas==2.2.3
pip==24.2
pytz==2025.2
numpy==2.2.6
python-dateutil==2.9.0.post0
six==1.17.0